\name{getApproxMat}
\alias{getApproxMat}
\title{getApproxMat}
\description{kernlab required for this
function to get low rank approximation with kernlab}
\usage{getApproxMat(geno, referenceMat, KernFun)}
\arguments{
  \item{geno}{m x n matrix}
  \item{referenceMat}{matrix de reference}
  \item{KernFun}{kernel function as in kernlab}
}

\value{return the approximate n x p matrix}


\author{Guillem Rigaill}



