\name{retour}
\alias{retour}
\title{retour}
\description{function to parse C code output and recover changes in i breaks}
\usage{retour(path, i)}
\arguments{
  \item{path}{matrix Mij of C output }
  \item{i}{line to consider}
}

\value{list of changes (internal use)}

\author{Guillem Rigaill}





