\name{SumsBlock_Gaussian}
\alias{SumsBlock_Gaussian}
\title{SumsBlock Gaussian}
\description{# compute all block sums in gram matrix
# from their we could get the cost matrix}
\usage{SumsBlock_Gaussian(geno, delta)}
\arguments{
  \item{geno}{
}
  \item{delta}{
}
}



\author{Guillem Rigaill}





